+++ src/app/login/page.tsx
@@ -7,7 +7,7 @@ import { useForm, SubmitHandler } from 'react-hook-form';
 
 import Link from 'next/link';
 
-import { useLogin } from '@/apis/queryHooks/Auth';
+import { useLogin } from '@/apis/queryHooks/useAuth/useLogin';
 import ErrorIcon from '@/assets/svg/error.svg';
 import GoggleIcon from '@/assets/svg/goggle.svg';
